# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "REST API | AETHER",
    "contact": {
      "name": "Aether",
      "url": "https://aetherz.xyz",
      "whatsapp": "+6285798045817"
    }
  },
  "paths": {
...
"/api/tiktok": {
  "get": {
    "tags": ["Downloader"],
    "summary": "Unduh Video TikTok",
    "description": "Endpoint untuk mengunduh video TikTok",
           "parameters": [
      {
        "name": "url",
        "in": "query",
        "description": "Url Tiktok Downloader",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "OK"
      }
    }
  }
},
...
    "/tiktok": {
      "get": {
        "tags": ["Downloader"],
        "summary": "Download video dari TikTok",
        "description": "Endpoint untuk download video dari TikTok",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "URL video di TikTok"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}